requirements.txt
elasticsearch==8.12.0
python-dotenv==1.0.0
requests==2.31.0
faker==22.6.0
pyyaml==6.0.1

google-generativeai==0.3.2
flask==3.0.0